+++ WebContent/WEB-INF/jsp/admin-list-condition.jsp
@@ -139,7 +139,7 @@
         <div class="container">
             <ol class="breadcrumb">
                 <li><a href="index-google-map-fullscreen">Home</a></li>
-                <li class="active">Admin Management</li>
+                <li class="active">List Condition</li>
             </ol>
         </div>
         <!-- end Breadcrumb -->
@@ -207,7 +207,7 @@
 								                                        <td class="actions">
 								                                            <a href="seacch-rate-week?conId=${val.conditionId }" class="edit"><i class="fa fa-pencil"></i>Rate</a>
 								                                        </td>
-								                                        <td class="actions"><a href="admin-delete-condition?conditionid=${val.conditionId }" class="edit"><i 
+								                                        <td class="actions"><a href="admin-delete-condition?conditionid=${val.conditionId }" class="edit" onclick="return confirm('You need delete ??');"><i 
 																					class="fa fa-trash-o" aria-hidden="true"></i>Delete</a>
 																		</td>
 								                                    </tr>
@@ -226,7 +226,7 @@
 								                                        <td class="actions">
 								                                            <a href="update-store?storeid=${val.conditionId }" class="edit"><i class="fa fa-pencil"></i>Edit</a>
 								                                        </td>
-								                                        <td class="actions"><a href="admin-delete-stores?storeid=${val.conditionId}" class="edit"><i 
+								                                        <td class="actions"><a href="admin-delete-condition?conditionid=${val.conditionId}" class="edit" onclick="return confirm('You need delete ??');"><i 
 																					class="fa fa-trash-o" aria-hidden="true"></i>Delete</a>
 																		</td>
 								                                    </tr>
